+++ src/entity/mod.rs
@@ -122,15 +122,15 @@ macro_rules! entity {
 }
 
 /// Defines the type of an entity containing the provided components.
-/// 
+///
 /// # Example
 /// ``` rust
 /// use brood::Entity;
-/// 
+///
 /// // Define components `Foo` and `Bar`.
 /// struct Foo(u16);
 /// struct Bar(f32);
-/// 
+///
 /// // Define the type for an entity containing the components `Foo` and `Bar`.
 /// type Entity = Entity!(Foo, Bar);
 /// ```